+++ android/src/main/kotlin/com/bruno/native_admob_flutter/native/NativeAd.kt
@@ -181,7 +181,9 @@ class NativeAdPlatformView(context: Context, data: Map<String?, Any?>?) : Platfo
 
         (data["backgroundColor"] as? String)?.let { shape.setColor(Color.parseColor(it)) }
 
-        view.background = shape
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+            view.background = shape
+        }
 
         // bounds
 
@@ -335,6 +337,5 @@ fun TextView.applyText(data: Map<*, *>) {
     (data["minLines"] as? Int?)?.let { view.minLines = it }
     (data["bold"] as? Boolean)?.let { if (it) view.setTypeface(view.typeface, Typeface.BOLD) }
     (data["text"] as? String)?.let { view.text = it }
-    view.lineHeight = 1
     view.ellipsize = TextUtils.TruncateAt.END
 }