+++ lib/src/app_open/app_open.dart
@@ -147,6 +147,9 @@ class AppOpenAd extends LoadShowAd<FullScreenAdEvent> {
 
     /// Whether non-personalized ads should be enabled
     bool? nonPersonalizedAds,
+
+    /// The keywords of the ad
+    List<String> keywords = const [],
   }) async {
     ensureAdNotDisposed();
     assertMobileAdsIsInitialized();
@@ -179,6 +182,7 @@ class AppOpenAd extends LoadShowAd<FullScreenAdEvent> {
           MobileAds.appOpenAdTestUnitId,
       'orientation': orientation,
       'nonPersonalizedAds': nonPersonalizedAds ?? this.nonPersonalizedAds,
+      'keywords': keywords,
     }).timeout(
       timeout ?? this.loadTimeout,
       onTimeout: () {