+++ lib/src/native/utils.dart
@@ -3,7 +3,7 @@ import 'layout_builder/layout_builder.dart';
 
 /// Default banner ad layout
 ///
-/// ![adBannerLayoutBuilder screenshot](https://github.com/bdlukaa/native_admob_flutter/blob/master/screenshots/native/banner_size_ad.png?raw=true)
+/// ![Banner Layout Builder Preview](https://github.com/bdlukaa/native_admob_flutter/blob/master/screenshots/native/banner_size_ad.png?raw=true)
 ///
 /// For more info, read the [documentation](https://github.com/bdlukaa/native_admob_flutter/wiki/Creating-a-native-ad#creating-a-layout-builder)
 AdLayoutBuilder get adBannerLayoutBuilder => (ratingBar, media, icon, headline,
@@ -35,6 +35,89 @@ AdLayoutBuilder get adBannerLayoutBuilder => (ratingBar, media, icon, headline,
       );
     };
 
+/// The small template is ideal for ListView, or any time
+/// you need a long rectangular ad view. For instance you
+/// could use it for in-feed ads.
+///
+/// The default height for this template is 115
+///
+/// ![Small Template Preview](https://developers.google.com/admob/images/android_small_template.png)
+AdLayoutBuilder get smallAdTemplateLayoutBuilder {
+  return (ratingBar, media, icon, headline, advertiser, body, price, store,
+      attribution, button) {
+    return AdLinearLayout(
+      decoration: AdDecoration(backgroundColor: Colors.white),
+      width: MATCH_PARENT,
+      height: MATCH_PARENT,
+      gravity: LayoutGravity.center_vertical,
+      padding: EdgeInsets.all(8.0),
+      children: [
+        attribution,
+        AdLinearLayout(
+          margin: EdgeInsets.only(top: 6.0),
+          orientation: HORIZONTAL,
+          children: [
+            icon,
+            AdExpanded(
+              flex: 2,
+              child: AdLinearLayout(
+                width: WRAP_CONTENT,
+                margin: EdgeInsets.symmetric(horizontal: 4),
+                children: [headline, body, advertiser],
+              ),
+            ),
+            AdExpanded(flex: 3, child: button),
+          ],
+        ),
+      ],
+    );
+  };
+}
+
+/// The medium template is meant to be a one-half to three-quarter
+/// page view but can also be used in feeds. It is good for landing
+/// pages or splash pages.
+///
+/// Feel free to experiment with placement. Of course, you can also
+/// change the source code to suit your requirements.
+///
+/// The default height for this template is 320
+///
+/// ![Medium Template Preview](https://developers.google.com/admob/images/android_medium_template.png)
+AdLayoutBuilder get mediumAdTemplateLayoutBuilder {
+  return (ratingBar, media, icon, headline, advertiser, body, price, store,
+      attribution, button) {
+    return AdLinearLayout(
+      decoration: AdDecoration(backgroundColor: Colors.white),
+      width: MATCH_PARENT,
+      height: MATCH_PARENT,
+      gravity: LayoutGravity.center_vertical,
+      padding: EdgeInsets.all(8.0),
+      children: [
+        attribution,
+        AdLinearLayout(
+          padding: EdgeInsets.only(top: 6.0),
+          height: WRAP_CONTENT,
+          orientation: HORIZONTAL,
+          children: [
+            icon,
+            AdExpanded(
+              flex: 2,
+              child: AdLinearLayout(
+                width: WRAP_CONTENT,
+                margin: EdgeInsets.symmetric(horizontal: 4),
+                children: [headline, body, advertiser],
+              ),
+            ),
+          ],
+        ),
+        media,
+        button
+      ],
+    );
+  };
+}
+
 extension colorExtension on Color {
   String toHex([bool hashtag = true]) {
     if (this == Colors.transparent) return '#00ff0000';
@@ -58,7 +141,7 @@ class AdBorderRadius {
   /// The bottom-left radius
   final double? bottomLeft;
 
-  AdBorderRadius({
+  const AdBorderRadius({
     this.topLeft,
     this.topRight,
     this.bottomLeft,