+++ lib/src/rewarded_interstitial/rewarded_interstitial.dart
@@ -161,6 +161,9 @@ class RewardedInterstitialAd extends LoadShowAd<RewardedAdEvent> {
 
     /// Whether non-personalized ads should be enabled
     bool? nonPersonalizedAds,
+
+    /// The keywords of the ad
+    List<String> keywords = const [],
   }) async {
     ensureAdNotDisposed();
     assertMobileAdsIsInitialized();
@@ -171,6 +174,7 @@ class RewardedInterstitialAd extends LoadShowAd<RewardedAdEvent> {
           MobileAds.rewardedAdUnitId ??
           MobileAds.rewardedAdTestUnitId,
       'nonPersonalizedAds': nonPersonalizedAds ?? this.nonPersonalizedAds,
+      'keywords': keywords,
     }).timeout(
       timeout ?? this.loadTimeout,
       onTimeout: () {